Split walk-in card tilt system
